requirements.txt
fastapi
uvicorn[standard]
gunicorn
motor
pymongo
python-dotenv
sendgrid
redis
celery
jinja2
python-multipart
requests
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
pydantic-settings
httpx
email-validator